verfiy api is changed.

+++ FreeFallTicketing.Application/PaymentServices/IPaymentService.cs
@@ -3,7 +3,7 @@
     public interface IPaymentService
     {
         Task<string> Checkout(double amount);
-        Task<int> Verify(string authority, double amount);
+        Task<ulong> Verify(string authority, double amount);
     }
 
     public interface IZarinpalPaymentService : IPaymentService { }
